requirements.txt
pandas
mysql-connector-python==9.4.0
python-dotenv==1.1.1
